a function that returns the sum of all the data (n) of a dlistint_t linked list

+++ 0x17-doubly_linked_lists/6-sum_dlistint.c
@@ -7,7 +7,7 @@
  * Return: number of nodes
  */
 
-size_t sum_dlistint(dlistint_t *head)
+int sum_dlistint(dlistint_t *head)
 {
 	size_t nodes = 0;
